+++ lingua_franca/lang/parse_fa.py
@@ -13,14 +13,11 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
-import json
-from datetime import timedelta
+from datetime import datetime, timedelta
 
-from lingua_franca.internal import resolve_resource_file
 from lingua_franca.lang.common_data_fa import (_FARSI_BIG, _FARSI_HUNDREDS,
                                                _FARSI_ONES, _FARSI_TENS,
                                                _FORMAL_VARIANT)
-from lingua_franca.lang.parse_common import Normalizer
 from lingua_franca.time import now_local
 
 
@@ -31,6 +28,7 @@ def _is_number(s):
     except ValueError:
         return False
 
+
 def _parse_sentence(text):
     for key, value in _FORMAL_VARIANT.items():
         text = text.replace(key, value)
@@ -39,8 +37,8 @@ def _parse_sentence(text):
     current_number = 0
     current_words = []
     s = 0
-    step = 10
     mode = 'init'
+
     def finish_num():
         nonlocal current_number
         nonlocal s
@@ -54,13 +52,14 @@ def _parse_sentence(text):
         current_number = 0
         current_words = []
         mode = 'init'
+
     for x in ar:
         if x == "و":
             if mode == 'num_ten' or mode == 'num_hundred' or mode == 'num_one':
                 mode += '_va'
                 current_words.append(x)
             elif mode == 'num':
-                current_words.append(x)    
+                current_words.append(x)
             else:
                 finish_num()
                 result.append(x)
@@ -71,7 +70,7 @@ def _parse_sentence(text):
         elif x in _FARSI_ONES:
             t = _FARSI_ONES.index(x)
             if mode != 'init' and mode != 'num_hundred_va' and mode != 'num':
-                if not(t < 10 and mode == 'num_ten_va'):
+                if not (t < 10 and mode == 'num_ten_va'):
                     finish_num()
             current_words.append(x)
             s += t
@@ -80,20 +79,20 @@ def _parse_sentence(text):
             if mode != 'init' and mode != 'num_hundred_va' and mode != 'num':
                 finish_num()
             current_words.append(x)
-            s += _FARSI_TENS.index(x)*10
+            s += _FARSI_TENS.index(x) * 10
             mode = 'num_ten'
         elif x in _FARSI_HUNDREDS:
             if mode != 'init' and mode != 'num':
                 finish_num()
             current_words.append(x)
-            s += _FARSI_HUNDREDS.index(x)*100
+            s += _FARSI_HUNDREDS.index(x) * 100
             mode = 'num_hundred'
         elif x in _FARSI_BIG:
             current_words.append(x)
             d = _FARSI_BIG.index(x)
             if mode == 'init' and d == 1:
                 s = 1
-            s *= 10**(3*d)
+            s *= 10 ** (3 * d)
             current_number += s
             s = 0
             mode = 'num'
@@ -119,6 +118,7 @@ _date_units = {
     'روز': timedelta(days=1),
     'هفته': timedelta(weeks=1),
 }
+
 
 def extract_duration_fa(text):
     """
@@ -208,9 +208,8 @@ def extract_datetime_fa(text, anchorDate=None, default_time=None):
         .replace('سه شنبه', 'سهشنبه') \
         .replace('چهار شنبه', 'چهارشنبه') \
         .replace('پنج شنبه', 'پنجشنبه') \
-        .replace('بعد از ظهر', 'بعدازظهر') \
-        
-        
+        .replace('بعد از ظهر', 'بعدازظهر')
+
     if not anchorDate:
         anchorDate = now_local()
     today = anchorDate.replace(hour=0, minute=0, second=0, microsecond=0)
@@ -225,11 +224,11 @@ def extract_datetime_fa(text, anchorDate=None, default_time=None):
         'یکشنبه',
     ]
     daysDict = {
-        'پریروز': today + timedelta(days= -2),
-        'دیروز': today + timedelta(days= -1),
+        'پریروز': today + timedelta(days=-2),
+        'دیروز': today + timedelta(days=-1),
         'امروز': today,
-        'فردا': today + timedelta(days= 1),
-        'پسفردا': today + timedelta(days= 2),
+        'فردا': today + timedelta(days=1),
+        'پسفردا': today + timedelta(days=2),
     }
     timesDict = {
         'صبح': timedelta(hours=8),
@@ -306,34 +305,6 @@ def extract_datetime_fa(text, anchorDate=None, default_time=None):
             number_seen = None
         remainder.append(x)
     return (result, " ".join(remainder))
-
-def is_fractional_fa(input_str, short_scale=True):
-    """
-    This function takes the given text and checks if it is a fraction.
-
-    Args:
-        input_str (str): the string to check if fractional
-        short_scale (bool): use short scale if True, long scale if False
-    Returns:
-        (bool) or (float): False if not a fraction, otherwise the fraction
-
-    """
-    if input_str.endswith('s', -1):
-        input_str = input_str[:len(input_str) - 1]  # e.g. "fifths"
-
-    fracts = {"whole": 1, "half": 2, "halve": 2, "quarter": 4}
-    if short_scale:
-        for num in _SHORT_ORDINAL_FA:
-            if num > 2:
-                fracts[_SHORT_ORDINAL_FA[num]] = num
-    else:
-        for num in _LONG_ORDINAL_FA:
-            if num > 2:
-                fracts[_LONG_ORDINAL_FA[num]] = num
-
-    if input_str.lower() in fracts:
-        return 1.0 / fracts[input_str.lower()]
-    return False
 
 
 def extract_numbers_fa(text, short_scale=True, ordinals=False):
